requirements.txt
setuptools~=61.2.0
pyyaml~=6.0
requests~=2.28.0
websockets~=10.3